refactor package for MongoAdditionalMappingContributor and its testing case
